--- handlers/src/alarm-config/utils/index.mts
@@ -1,12 +1,0 @@
-export {
-  EC2getCpuQuery,
-  EC2getMemoryQuery,
-  EC2getStorageQuery,
-  batchPromRulesDeletion,
-  batchUpdatePromRules,
-  makeSignedRequest,
-  queryPrometheusForService,
-  describeNamespace,
-  managePromNamespaceAlarms,
-  deletePromRulesForService,
-} from './prometheus/index.mjs';
