--- handlers/src/alarm-config/utils/prometheus/index.mts
@@ -1,14 +1,0 @@
-export {
-  EC2getCpuQuery,
-  EC2getMemoryQuery,
-  EC2getStorageQuery,
-} from './prometheus-queries.mjs';
-export {
-  batchPromRulesDeletion,
-  batchUpdatePromRules,
-  makeSignedRequest,
-  queryPrometheusForService,
-  describeNamespace,
-  managePromNamespaceAlarms,
-  deletePromRulesForService,
-} from './prometheus-tools.mjs';
