+++ handlers/src/alarm-configs/cloudfront-configs.mts
@@ -1,46 +1,46 @@
 /**
- * @fileoverview TRANSIT_GATEWAY alarm configuration definitions.
+ * @fileoverview CLOUDFRONT alarm configuration definitions.
  *
- * This file contains the default configurations for all supported TRANSIT_GATEWAY CloudWatch alarms managed by AutoAlarm.
+ * This file contains the default configurations for all supported CLOUDFRONT CloudWatch alarms managed by AutoAlarm.
  *
  * @requires
  * - Approval from Owners Team lead and consultation before adding new alarms
  * - Anomaly Alarms can only use the following comparison operators: GreaterThanUpperThreshold, LessThanLowerOrGreaterThanUpperThreshold, LessThanLowerThreshold
  *
- * @Owners NETWORKING
+ * @Owners HARMONY-DEVOPS
  */
 
-import {MetricAlarmConfig} from '../../types/index.mjs';
+import {MetricAlarmConfig} from '../types/index.mjs';
 import {ComparisonOperator} from '@aws-sdk/client-cloudwatch';
 import {TreatMissingData} from 'aws-cdk-lib/aws-cloudwatch';
 
 /**
- * _TRANSIT_GATEWAY alarm configuration definitions.
+ * _CLOUDFRONT alarm configuration definitions.
  * Implements the {@link MetricAlarmConfig} interface.
  * Used to map a tag key to a CloudWatch metric name and namespace to default alarm configurations {@link MetricAlarmOptions}.
  */
-export const TRANSIT_GATEWAY_CONFIGS: MetricAlarmConfig[] = [
+export const CLOUDFRONT_CONFIGS: MetricAlarmConfig[] = [
   {
-    tagKey: 'bytesin',
-    metricName: 'BytesIn',
-    metricNamespace: 'AWS/TransitGateway',
+    tagKey: '4xx-errors',
+    metricName: '4xxErrorRate',
+    metricNamespace: 'AWS/CloudFront',
     defaultCreate: true,
     anomaly: false,
     defaults: {
-      warningThreshold: 187500000000,
-      criticalThreshold: 225000000000,
+      warningThreshold: 5, // Adjust threshold based on your needs
+      criticalThreshold: 10, // Adjust threshold based on your needs
       period: 300,
       evaluationPeriods: 1,
-      statistic: 'Sum',
+      statistic: 'Average',
       dataPointsToAlarm: 1,
       comparisonOperator: ComparisonOperator.GreaterThanThreshold,
       missingDataTreatment: TreatMissingData.IGNORE,
     },
   },
   {
-    tagKey: 'bytesin-anomaly',
-    metricName: 'BytesIn',
-    metricNamespace: 'AWS/TransitGateway',
+    tagKey: '4xx-errors-anomaly',
+    metricName: '4xxErrorRate',
+    metricNamespace: 'AWS/CloudFront',
     defaultCreate: false,
     anomaly: true,
     defaults: {
@@ -55,26 +55,26 @@ export const TRANSIT_GATEWAY_CONFIGS: MetricAlarmConfig[] = [
     },
   },
   {
-    tagKey: 'bytesout',
-    metricName: 'BytesOut',
-    metricNamespace: 'AWS/TransitGateway',
+    tagKey: '5xx-errors',
+    metricName: '5xxErrorRate',
+    metricNamespace: 'AWS/CloudFront',
     defaultCreate: true,
     anomaly: false,
     defaults: {
-      warningThreshold: 187500000000,
-      criticalThreshold: 225000000000,
+      warningThreshold: 1,
+      criticalThreshold: 5,
       period: 300,
       evaluationPeriods: 1,
-      statistic: 'Sum',
+      statistic: 'Average',
       dataPointsToAlarm: 1,
       comparisonOperator: 'GreaterThanThreshold',
       missingDataTreatment: 'ignore',
     },
   },
   {
-    tagKey: 'bytesout-anomaly',
-    metricName: 'BytesOut',
-    metricNamespace: 'AWS/TransitGateway',
+    tagKey: '5xx-errors-anomaly',
+    metricName: '5xxErrorRate',
+    metricNamespace: 'AWS/CloudFront',
     defaultCreate: false,
     anomaly: true,
     defaults: {
