+++ handlers/src/alarm-configs/index.mts
@@ -1,0 +1,15 @@
+/**
+ * Barrel file to export all alarm configuration modules.
+ */
+export {ALB_CONFIGS} from './alb-configs.mjs';
+export {CLOUDFRONT_CONFIGS} from './cloudfront-configs.mjs';
+export {EC2_CONFIGS} from './ec2-configs.mjs';
+export {OPENSEARCH_CONFIGS} from './opensearch-configs.mjs';
+export {RDS_CLUSTER_CONFIGS} from './rds-cluster-configs.mjs';
+export {RDS_CONFIGS} from './rds-configs.mjs';
+export {ROUTE53_RESOLVER_CONFIGS} from './route53-resolver-configs.mjs';
+export {SQS_CONFIGS} from './sqs-configs.mjs';
+export {STEP_FUNCTION_CONFIGS} from './step-function-configs.mjs';
+export {TARGET_GROUP_CONFIGS} from './target-group-configs.mjs';
+export {TRANSIT_GATEWAY_CONFIGS} from './transit-gateway-configs.mjs';
+export {VPN_CONFIGS} from './vpn-configs.mjs';
