+++ handlers/src/alarm-configs/sqs-configs.mts
@@ -10,10 +10,9 @@
  * @Owners HARMONY-DEVOPS
  */
 
-import {MetricAlarmConfig} from '../../types/index.mjs';
+import {MetricAlarmConfig} from '../types/index.mjs';
 import {ComparisonOperator} from '@aws-sdk/client-cloudwatch';
 import {TreatMissingData} from 'aws-cdk-lib/aws-cloudwatch';
-
 
 /**
  * _SQS alarm configuration definitions.
