+++ handlers/src/alarm-configs/step-function-configs.mts
@@ -1,7 +1,7 @@
 /**
- * @fileoverview ROUTE53_RESOLVER alarm configuration definitions.
+ * @fileoverview STEP_FUNCTION alarm configuration definitions.
  *
- * This file contains the default configurations for all supported ROUTE53_RESOLVER CloudWatch alarms managed by AutoAlarm.
+ * This file contains the default configurations for all supported STEP_FUNCTION CloudWatch alarms managed by AutoAlarm.
  *
  * @requires
  * - Approval from Owners Team lead and consultation before adding new alarms
@@ -10,26 +10,26 @@
  * @Owners HARMONY-DEVOPS
  */
 
-import {MetricAlarmConfig} from '../../types/index.mjs';
+import {MetricAlarmConfig} from '../types/index.mjs';
 import {ComparisonOperator} from '@aws-sdk/client-cloudwatch';
 import {TreatMissingData} from 'aws-cdk-lib/aws-cloudwatch';
 
 /**
- * _ROUTE53_RESOLVER alarm configuration definitions.
+ * _STEP_FUNCTION alarm configuration definitions.
  * Implements the {@link MetricAlarmConfig} interface.
  * Used to map a tag key to a CloudWatch metric name and namespace to default alarm configurations {@link MetricAlarmOptions}.
  */
-export const ROUTE53_RESOLVER: MetricAlarmConfig[] = [
+export const STEP_FUNCTION_CONFIGS: MetricAlarmConfig[] = [
   {
-    tagKey: 'inbound-query-volume',
-    metricName: 'InboundQueryVolume',
-    metricNamespace: 'AWS/Route53Resolver',
+    tagKey: 'executions-failed',
+    metricName: 'ExecutionsFailed',
+    metricNamespace: 'AWS/States',
     defaultCreate: true,
     anomaly: false,
     defaults: {
-      warningThreshold: 1500000,
-      criticalThreshold: 2000000,
-      period: 300,
+      warningThreshold: null,
+      criticalThreshold: 1,
+      period: 60,
       evaluationPeriods: 1,
       statistic: 'Sum',
       dataPointsToAlarm: 1,
@@ -38,15 +38,15 @@ export const ROUTE53_RESOLVER: MetricAlarmConfig[] = [
     },
   },
   {
-    tagKey: 'inbound-query-volume-anomaly',
-    metricName: 'InboundQueryVolume',
-    metricNamespace: 'AWS/Route53Resolver',
+    tagKey: 'executions-failed-anomaly',
+    metricName: 'ExecutionsFailed',
+    metricNamespace: 'AWS/States',
     defaultCreate: false,
     anomaly: true,
     defaults: {
       warningThreshold: null,
       criticalThreshold: null,
-      period: 300,
+      period: 60,
       evaluationPeriods: 1,
       statistic: 'Average',
       dataPointsToAlarm: 1,
@@ -55,15 +55,15 @@ export const ROUTE53_RESOLVER: MetricAlarmConfig[] = [
     },
   },
   {
-    tagKey: 'outbound-query-volume',
-    metricName: 'OutboundQueryAggregateVolume',
-    metricNamespace: 'AWS/Route53Resolver',
+    tagKey: 'executions-timed-out',
+    metricName: 'ExecutionsTimedOut',
+    metricNamespace: 'AWS/States',
     defaultCreate: true,
     anomaly: false,
     defaults: {
-      warningThreshold: 1500000,
-      criticalThreshold: 2000000,
-      period: 300,
+      warningThreshold: null,
+      criticalThreshold: 1,
+      period: 60,
       evaluationPeriods: 1,
       statistic: 'Sum',
       dataPointsToAlarm: 1,
@@ -72,15 +72,15 @@ export const ROUTE53_RESOLVER: MetricAlarmConfig[] = [
     },
   },
   {
-    tagKey: 'outbound-query-volume-anomaly',
-    metricName: 'OutboundQueryAggregateVolume',
-    metricNamespace: 'AWS/Route53Resolver',
+    tagKey: 'executions-timed-out-anomaly',
+    metricName: 'ExecutionsTimedOut',
+    metricNamespace: 'AWS/States',
     defaultCreate: false,
     anomaly: true,
     defaults: {
       warningThreshold: null,
       criticalThreshold: null,
-      period: 300,
+      period: 60,
       evaluationPeriods: 1,
       statistic: 'Average',
       dataPointsToAlarm: 1,
