+++ handlers/src/alarm-configs/transit-gateway-configs.mts
@@ -1,35 +1,35 @@
 /**
- * @fileoverview STEP_FUNCTION alarm configuration definitions.
+ * @fileoverview TRANSIT_GATEWAY alarm configuration definitions.
  *
- * This file contains the default configurations for all supported STEP_FUNCTION CloudWatch alarms managed by AutoAlarm.
+ * This file contains the default configurations for all supported TRANSIT_GATEWAY CloudWatch alarms managed by AutoAlarm.
  *
  * @requires
  * - Approval from Owners Team lead and consultation before adding new alarms
  * - Anomaly Alarms can only use the following comparison operators: GreaterThanUpperThreshold, LessThanLowerOrGreaterThanUpperThreshold, LessThanLowerThreshold
  *
- * @Owners HARMONY-DEVOPS
+ * @Owners NETWORKING
  */
 
-import {MetricAlarmConfig} from '../../types/index.mjs';
+import {MetricAlarmConfig} from '../types/index.mjs';
 import {ComparisonOperator} from '@aws-sdk/client-cloudwatch';
 import {TreatMissingData} from 'aws-cdk-lib/aws-cloudwatch';
 
 /**
- * _STEP_FUNCTION alarm configuration definitions.
+ * _TRANSIT_GATEWAY alarm configuration definitions.
  * Implements the {@link MetricAlarmConfig} interface.
  * Used to map a tag key to a CloudWatch metric name and namespace to default alarm configurations {@link MetricAlarmOptions}.
  */
-export const STEP_FUNCTION_CONFIGS: MetricAlarmConfig[] = [
+export const TRANSIT_GATEWAY_CONFIGS: MetricAlarmConfig[] = [
   {
-    tagKey: 'executions-failed',
-    metricName: 'ExecutionsFailed',
-    metricNamespace: 'AWS/States',
+    tagKey: 'bytesin',
+    metricName: 'BytesIn',
+    metricNamespace: 'AWS/TransitGateway',
     defaultCreate: true,
     anomaly: false,
     defaults: {
-      warningThreshold: null,
-      criticalThreshold: 1,
-      period: 60,
+      warningThreshold: 187500000000,
+      criticalThreshold: 225000000000,
+      period: 300,
       evaluationPeriods: 1,
       statistic: 'Sum',
       dataPointsToAlarm: 1,
@@ -38,15 +38,15 @@ export const STEP_FUNCTION_CONFIGS: MetricAlarmConfig[] = [
     },
   },
   {
-    tagKey: 'executions-failed-anomaly',
-    metricName: 'ExecutionsFailed',
-    metricNamespace: 'AWS/States',
+    tagKey: 'bytesin-anomaly',
+    metricName: 'BytesIn',
+    metricNamespace: 'AWS/TransitGateway',
     defaultCreate: false,
     anomaly: true,
     defaults: {
       warningThreshold: null,
       criticalThreshold: null,
-      period: 60,
+      period: 300,
       evaluationPeriods: 1,
       statistic: 'Average',
       dataPointsToAlarm: 1,
@@ -55,15 +55,15 @@ export const STEP_FUNCTION_CONFIGS: MetricAlarmConfig[] = [
     },
   },
   {
-    tagKey: 'executions-timed-out',
-    metricName: 'ExecutionsTimedOut',
-    metricNamespace: 'AWS/States',
+    tagKey: 'bytesout',
+    metricName: 'BytesOut',
+    metricNamespace: 'AWS/TransitGateway',
     defaultCreate: true,
     anomaly: false,
     defaults: {
-      warningThreshold: null,
-      criticalThreshold: 1,
-      period: 60,
+      warningThreshold: 187500000000,
+      criticalThreshold: 225000000000,
+      period: 300,
       evaluationPeriods: 1,
       statistic: 'Sum',
       dataPointsToAlarm: 1,
@@ -72,15 +72,15 @@ export const STEP_FUNCTION_CONFIGS: MetricAlarmConfig[] = [
     },
   },
   {
-    tagKey: 'executions-timed-out-anomaly',
-    metricName: 'ExecutionsTimedOut',
-    metricNamespace: 'AWS/States',
+    tagKey: 'bytesout-anomaly',
+    metricName: 'BytesOut',
+    metricNamespace: 'AWS/TransitGateway',
     defaultCreate: false,
     anomaly: true,
     defaults: {
       warningThreshold: null,
       criticalThreshold: null,
-      period: 60,
+      period: 300,
       evaluationPeriods: 1,
       statistic: 'Average',
       dataPointsToAlarm: 1,
