+++ handlers/src/types/index.mts
@@ -1,11 +1,14 @@
+/**
+ * Barrel file to export all types and enums
+ */
 // Import and re-export from alarm-config-types.mjs
 export {
   ValidExtendedStat,
   ValidStatistic,
   MissingDataTreatment,
   MetricAlarmOptions,
-  MetricAlarmConfig
-} from "./alarm-config-types.mjs";
+  MetricAlarmConfig,
+} from './alarm-config-types.mjs';
 
 // Import and re-export from enums.mjs
 export {
@@ -14,8 +17,8 @@ export {
   ValidAlbEvent,
   ValidTargetGroupEvent,
   ValidSqsEvent,
-  ValidOpenSearchState
-} from "./enums.mjs";
+  ValidOpenSearchState,
+} from './enums.mjs';
 
 // Import and re-export from module-types.mjs
 export {
@@ -24,8 +27,8 @@ export {
   Tag,
   PathMetrics,
   LoadBalancerIdentifiers,
-  AnomalyAlarmProps
-} from "./module-types.mjs";
+  AnomalyAlarmProps,
+} from './module-types.mjs';
 
 // Import and re-export from prometheus-types.mjs
 export {
@@ -33,5 +36,5 @@ export {
   AMPRule,
   RuleGroup,
   PrometheusAlarmConfigArray,
-  NamespaceDetails
-} from "./prometheus-types.mjs";
+  NamespaceDetails,
+} from './prometheus-types.mjs';
