+++ handlers/tsconfig.json
@@ -15,7 +15,7 @@
     "noImplicitReturns": true,
     "pretty": true,
     "sourceMap": true,
-    "skipLibCheck": true,
+    "skipLibCheck": true
   },
   "include": ["src"],
   "exclude": ["node_modules"]
