+++ package.json
@@ -21,7 +21,7 @@
     "clean-cdk-lib": "cd cdk/lib ; find . -type f \\( -name '*.d.ts' -o -name '*.js' \\) -delete",
     "clean-cdk-bin": "cd cdk/bin ; find . -type f \\( -name '*.d.ts' -o -name '*.js' \\) -delete",
     "clean": "concurrently \"pnpm:clean-handlers\" \"pnpm:clean-cdk-lib\" \"pnpm:clean-cdk-bin\" -c blue,green,yellow",
-    "create-cw-configs": "tsx handlers/src/alarm-config/cloudwatch-alarms/_template-configs.mts",
+    "create-cw-configs": "tsx handlers/src/alarm-configs/cloudwatch-alarms/_template-configs.mts",
     "watch-cdk": "cd cdk && cdk watch",
     "deploy-cdk": "cd cdk && cdk deploy",
     "destroy-cdk": "cd cdk && cdk destroy",
